+++ pywps/app/__init__.py
@@ -7,7 +7,6 @@
 
 from pywps.app.Process import Process  # noqa: F401
 from pywps.app.Service import Service  # noqa: F401
-from pywps.app.WPSResponse import WPSResponse  # noqa: F401
 from pywps.app.WPSRequest import WPSRequest  # noqa: F401
 from pywps.app.WPSRequest import get_inputs_from_xml  # noqa: F401
 from pywps.app.WPSRequest import get_output_from_xml  # noqa: F401
